#step3_1 [refactor] package 구조 변경
util에서 여러가지 인터페이스와 클래스가 섞일수가 있어서 하위에 각각의 package를 하나 더 생성
