+++ src/main/java/racingcar/util/drivingStrategy/DrivingStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.util.drivingStrategy;
 
 public interface DrivingStrategy {
     int throwDice();
