+++ src/main/java/racingcar/util/drivingStrategy/RandomStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.util.drivingStrategy;
 
 import java.util.Random;
