+++ src/test/java/racingcar/util/drivingStrategy/MockDrivingStrategy.java
@@ -1,6 +1,4 @@
-package racingcar.mocks;
-
-import racingcar.util.DrivingStrategy;
+package racingcar.util.drivingStrategy;
 
 /**
  *
