+++ src/test/java/racingcar/util/drivingStrategy/RandomStrategyTest.java
@@ -1,10 +1,8 @@
-package racingcar.mocks;
+package racingcar.util.drivingStrategy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
-import racingcar.util.DrivingStrategy;
-import racingcar.util.RandomStrategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
